feat: Introduce abstract resource access controller
